Allow GCN_LSTM to be saved and loaded (#1850)
This allows the `A` parameter to the underlying
`FixedAdjacencyGraphConvolution` layer to be `None`. If it's `None`, the
adjacency matrix weight is created with zeros. In the case of saving/loading
models, the weight is serialised during save and the `A` parameter is specified
as None in the configuration, thus, when the model is loaded, the weight is
created with zeros, but overwritten with the serialised weight.

Note: setting `A` to `None` shouldn't generally be used explicitly (only while
saving/loading models), and so this isn't documented.

Fixes #1681