Fix flaky test_normalize_adj in tests.core.test_utils (#1987)
Closes #1678

This PR fixes `test_normalize_adj` in `tests.core.test_utils` which was failing
in a flaky way.

The test was occasionally failing in two places:

- Comparing the top eigenvalue of symmetrically normalized adjacency matrix
  with 1. See, for example,
  https://buildkite.com/stellar/stellargraph-public/builds/3837

- Checking the sum of elements of non-symmetrically normalized adjacency
  matrix. See, for example,
  https://buildkite.com/stellar/stellargraph-public/builds/3967

The first point of failure seems to be the question of precision when
calculating eigenvalues. The current error threshold equal to 1e-7 is too
low. I ran 100000 iterations of the test and got 14963 occasions when the error
is > 1e-8 and 33 occasions when the error is > 1e-7. By setting the threshold
to 1e-5 we would make it virtually impossible for the test to fail.

The second point of failure is similar to
https://github.com/stellargraph/stellargraph/pull/1986. The sum of elements of
the normalized adjacency matrix should be compared to the number of nodes of
positive degree (each row contributes 1 except for those corresponding to
degree 0 nodes). For graphs generated with `example_graph_random` this is equal
to 5 most of the time, but not always - that's why the test was failing

I also remove unused import in `stellargraph/data/epgm.py`, which is not in the
dependencies and caused many CI jobs to fail - let me know if you would prefer
to keep it and just add to the dependencies.

## Tests

The following code runs without errors:
```
from tests.core.test_utils import example_graph_random, test_normalize_adj
for i in range(100000):
    random_graph = example_graph_random()
    test_normalize_adj(random_graph)
```